requirements.txt
pydantic~=2.10.6
pytest~=8.3.5
loguru~=0.7.3
tiktoken~=0.9.0
pyyaml~=6.0.2